Further improve ES2015 compliance
There are massive changes to the ESLint file, though I hopefully didn't
turn on anything *too* egregious.  Anyway, this commit makes the code as
ES2015-compliant as native Node.js will allow.  Nothing in here should
not be supported by Node 6.x, but I still recommend the field in
`package.json` for specifying Node 8+, which is in active LTS as of the
time of writing, in which case, the project may wish to consider going
to ES2016 or 2017 support instead to nab that exponent operator as well
as the object spread operator.

Further additions to the code include a beefed-up `.gitignore` file,
courtesy of https://gitignore.io, as well as a small `.eslintrc.json`
file in the testing area.  The latter allows for the removal of the
comment headers about `define` and `it`, as they are defined as part of
the mocha environment.  It inherits everything from the settings above
it, too, so it's able to be super compact.

Signed-off-by: Jonathan Sifuentes <jayands.dev@gmail.com>