Apply the changes mentioned
I have made the small changes necessary, as well as made one effective:
Travis will now actually run the code against Node 6, if it's to be
supported properly.  As soon as any configuration for ESLint is decided
upon, I will add it to the mainline code and clean up appropriately.

Signed-off-by: Jonathan Sifuentes <jayands.dev@gmail.com>