Commit for Issue 39
It's not done, but I have committed the code in an attempt to show what
this codebase should look like in my opinion.  The two tests that fail
do so because of timeouts in the `encoder.test.js` file.  That's the
main reason I haven't actually pushed this before now, but I have a
compelling reason to do so at the moment, namely that the branch will
exist.

I'm not sure where the code leak is though.  My next thought is to have
code coverage analysis and possibly static typing analysis.

Signed-off-by: Jonathan Sifuentes <jayands.dev@gmail.com>