A fully working test suite
I'm not sure about blindly using `flow` anymore; I had 99.8% coverage in
`tools` but broke the decoder's tests when I tried to get down to 0
errors in flow.

At any rate, the tests now work with a very impressive amount of code
coverage.  Additionally, I added nearly every output type for Rollup,
and it uses Node's built-ins, similar to Browserify, eliminating the
need for that one check for `TextDecoder` since it just uses Node's
`Buffer`.

With this commit, I'd even go so far as to say it's on the road to *at
least* 2.3, if not a full 3.0 release.  I'm not sure about 3.0 since I
didn't really add outward-facing features, but I fixed the test suite as
well as some of the subtler errors, so 2.3 definite contender.

Signed-off-by: Jonathan Sifuentes <jayands.dev@gmail.com>