Tailor the tests to match only one block
I don't know how it is with Mocha, but Jest keeps going with the emitter
tests.  So if it matches more than one thing, it's going to match them
against any assertions.  This was a source of much confusion for me over
the past two weeks, because I was under the impression that the tests
were absolutely correct.  The ones in value, in specific, weren't.  I
had implemented a solution, but it doesn't work for the older Nodes, so
I instead tailored the tests to match only one block based on the files
themselves, and removed that solution.

Hopefully *this* will pass in 6 and 8, otherwise I got nothing,
currently.

Signed-off-by: Jonathan Sifuentes <jayands.dev@gmail.com>