Cleanup and optimizations
This commit adds in a `babel.config.js`, supplanting the `.babelrc` file
in favor of babel's recommendation to libraries. Furthermore, it adds
more ignorable files to `.npmignore`, as these are either too localized
or should be acquired from the git repository instead. Lastly,
`.jshintrc` is removed since eslint is the preferred linting tool now.

I also decided to completely strip flow typing from most of the project,
but it _is_ on the agenda to be added as an issue.