[Misc] Add a yarnclean file
If `yarn` is used as the package manager, this file will enable the
removal of files that may seem extraneous to the packaging, that were
not covered by the package's `.npmignore` file, if they exist.  This is
run automatically at the time of `yarn install` or manually by running
`yarn autoclean`.

I elected to ignore the `README.md` files in `.yarnclean` since they may
also contain useful examples or references to API calls.

Signed-off-by: Jonathan Sifuentes <jayands.dev@gmail.com>