Convert pipeline tests to `unexpected`
The first file was the shortest to convert.  It also gives you a good
idea of how tests are written with unexpected.  If something fails,
however, it's equally friendly in its output.

The next few commits will just be this kind of thing,
with everything passing exactly as it should.

Signed-off-by: Jonathan Sifuentes <jayands.dev@gmail.com>