Decoder tests passing
I've injected the date functions for when the dates will be decoded, but
otherwise, they're the same tests.  There is an interesting block at the
bottom, however:  Instead of strictly deciding that the unknown tag is
exactly one way forever, `expect.it` enables subtesting of the key's
properties, e.g., if it's to be assumed that a key that we don't know
about is just going to be some obscure/custom WebM property, the test
will now still pass when that change is made to the source code.

Signed-off-by: Jonathan Sifuentes <jayands.dev@gmail.com>