Cleanup of Encoder and Tools
This push finally removes the dependency on `buffers`, which seemed to
be replaceable (in our use-case), with either simple `Buffer` instances
or `Array.from`.

Unfortunately, there are two now-known faulty tests that I could not
solve, so `assert` has to stay for now:

- EBML › Encoder › #cork and #uncork › should block flushing when corked
- EBML › Encoder › #cork and #uncork › should not block flushing when uncorked

They use `assert.ok` in an incorrect manner.