Encoder tests, mostly rewritten
I'll open an issue regarding the last two tests, but they are known to
be broken now. It's because of the fact that the buffer is flushed
and/or nulled out before I get a chance to check it, if I had to hazard
a guess.

At any rate, they are blocking the tests from passing, so I'm pushing
the tests as-is, but it's an issue to be addressed.