[Chore] Update tooling to 2022 (start)
The code is in significant need of updating.  BigInt was accepted,
TypeScript became the dominant typing system, and the way tools work
have been reworked.  I won't guarantee it'll be perfect, but I will try
to maintain this library, and if I can't anymore, pass it on to someone
who can.

Signed-off-by: Jonathan Sifuentes <jayands.dev@gmail.com>